backend/requirements.txt
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
cachetools==5.3.1
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
Django==4.0.4
django-jsonform==2.9.0
djangorestframework==3.13.1
executing==0.8.3
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.3
ipython==8.3.0
jedi==0.18.1
matplotlib-inline==0.1.3
oauthlib==3.2.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.29
protobuf==4.23.4
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.12.0
pyparsing==3.1.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
stack-data==0.2.0
traitlets==5.2.0
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
